Doc_Chatbot/requirements.txt
# ── Core framework ─────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LangChain ──────────────────────────────────────────────────────────────────
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16

# ── Vector store ───────────────────────────────────────────────────────────────
qdrant-client==1.17.0          

# ── HTTP client ────────────────────────────────────────────────────────────────
requests>=2.31.0

# ── Document parsing ───────────────────────────────────────────────────────────
pdfplumber>=0.10.0
python-docx>=1.1.0

# ── Numerics ───────────────────────────────────────────────────────────────────
numpy>=1.26.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
